vision_node/requirements.txt
opencv-python-headless>=4.8
paho-mqtt>=2.0
flask>=3.0
pyyaml>=6.0
numpy>=1.24

# RPi only — install via apt, NOT pip:
#   sudo apt install python3-picamera2
#
# SSOCR — install via apt or build from repo:
#   sudo apt install ssocr
#   # or: cd ssocr/ && make && sudo make install
